+++ docs/changelog.rst
@@ -22,6 +22,8 @@ Released on 2014-xx-xx.
   ``albums_sort_reverse``, ``medias_sort_attr``, ``medias_sort_reverse``.
 - New setting ``autorotate_images`` to disable autorotation of images, and warn
   about the incompatibility between autorotation and EXIF copy.
+- New settings ``ignore_directories`` and ``ignore_files`` to filter
+  directories and files with pattern matching.
 
 Version 0.6.0
 ~~~~~~~~~~~~~